requirements.txt
wget==3.2

PyYAML~=5.3.1